+++ src/constants/apiTimeframeOptions.ts
@@ -1,0 +1,6 @@
+export const apiTimeframeOptions = {
+  oneDay: 'one_day',
+  twoDays: 'two_days',
+  oneWeek: 'one_week',
+  oneMonth: 'one_month',
+}
