+++ src/pages/Pool/index.tsx
@@ -29,7 +29,7 @@ import { ToggleElement, ToggleWrapper } from 'components/Toggle/MultiToggle'
 import { formattedNum, formattedPercent, get2DayPercentChange, getPercentChange } from 'utils/dashboard'
 import { REWARDS_SELECTOR, STARKNET_REWARDS_API_URL } from 'constants/misc'
 import { HISTORICAL_GLOBAL_DATA } from 'graphql/data/queries'
-import { apiTimeframeOptions } from 'constants/dashboardApi'
+import { apiTimeframeOptions } from 'constants/apiTimeframeOptions'
 import { DEFAULT_CHAIN_ID, ETH_ADDRESS, NONFUNGIBLE_POOL_MANAGER_ADDRESS } from 'constants/tokens'
 import { providerInstance } from 'utils/getLibrary'
 import { cairo, hash, num, uint256 } from 'starknet'
